实现了基本的权限管理功能，数据库的增删，表的增删改查权限管理

+++ main.py
@@ -6,11 +6,19 @@
 import csv
 import sys
 from getpass import getpass
+from pathlib import Path
 import re
 import baseTool
 import tableTool
+import joinTool
+import privilege
 
+
+
+base_dir = Path(__file__).cwd()
+userfile = base_dir.joinpath("user.csv")
 current_database_name = None
+
 def show_usage():
     print("Error usage!")
     print("Usage:python ./main.py -u | --user [user] -p | --password")
@@ -18,9 +26,44 @@ def show_usage():
     print("Then input your password!")
 
 
-def handle_command(command):
-    print(command)
+
+
+def handle_command(username,command):
+    
     global current_database_name
+    # 审核一下权限
+    infos = None
+    with open(userfile,"r",newline="") as csvfile:
+        reader = csv.DictReader(csvfile,delimiter="|")
+        for item in reader:
+            if item["USERNAME"] == username:
+                infos = item
+                break
+    
+    database_table_list = infos["DATABASE_TABLE"].split(",")
+
+    di = {}
+    
+    li = [ database_table.split(".") for database_table in database_table_list]
+    print(li)
+    for i in li:
+        if i[0] not in di.keys():
+            di[i[0]] = []
+            di[i[0]].append(i[1])
+        else:
+            di[i[0]].append(i[1])
+
+    print(di)
+
+
+    is_create = int(infos["CREATE"])
+    is_drop = int(infos["DROP"])
+    is_insert = int(infos["INSERT"])
+    is_delete = int(infos["DELETE"])
+    is_update = int(infos["UPDATE"])
+    is_select = int(infos["SELECT"])
+    is_with_grant_option = int(infos["GRANTOPTION"])
+
     # 先列出所有的 匹配项
     is_create_database = re.compile(r"create database (\w+);",re.IGNORECASE).match(command)    
     is_drop_database = re.compile(r"drop database (\w+);",re.IGNORECASE).match(command)
@@ -45,14 +88,38 @@ def handle_command(command):
     is_select_from_order_by = re.compile(r"select ([\s\S]+) from (\w+) order by (\w+);",re.IGNORECASE).match(command)
     is_select_from_group_by = re.compile(r"select ([\s\S]+) from (\w+) group by ([\s\S]+);",re.IGNORECASE).match(command)
 
-    # CREATE DATABASE database-name 实现创建一个数据库的功能
-    if is_create_database:
+    # 实现连接查询
+    is_inner_join = re.compile(r"select ([\s\S]+) from (\w+) inner join (\w+) on ([\s\S]+);",re.IGNORECASE).match(command)
+    # 左边独有的数据 加上共有的
+    is_left_join = re.compile(r"select ([\s\S]+) from (\w+) left join (\w+) on ([\s\S]+);",re.IGNORECASE).match(command)
+    # 右边独有的数据，加上共有的
+    is_right_join = re.compile(r"select ([\s\S]+) from (\w+) right join (\w+) on ([\s\S]+);",re.IGNORECASE).match(command)
+
+
+
+    # 是否是设置用户，创建用户，设置权限的操作
+    # grant all privileges on *.* to jack@'localhost' identified by "jack" with grant option;
+    is_grant_privilege_on_user_host_password_is = re.compile(r"grant ([\s\S]+) on ([\s\S]+) to '(\w+)'@'([\s\S]+)' identified by '([\s\S]+)'([\s\S]*);",re.IGNORECASE).match(command)
+
+    if is_grant_privilege_on_user_host_password_is and is_with_grant_option:
+        chmod = is_grant_privilege_on_user_host_password_is.group(1).strip()
+        d_t = is_grant_privilege_on_user_host_password_is.group(2).strip()
+        user = is_grant_privilege_on_user_host_password_is.group(3).strip()
+        host = is_grant_privilege_on_user_host_password_is.group(4).strip()
+        pw = is_grant_privilege_on_user_host_password_is.group(5).strip()
+        is_wgo = is_grant_privilege_on_user_host_password_is.group(6).strip()
+        privilege.handle(chmod,d_t,user,host,pw,is_wgo)
+        return
+
+    # 加入权限信息
+    # CREATE DATABASE database-name 实现创建一个数据库的功能  
+    if is_create_database and is_create:
         database_name = is_create_database.group(1)
         baseTool.create_database(database_name)
         return
 
     # drop database dbname   实现删除一个数据库的功能
-    if is_drop_database:
+    if is_drop_database and is_drop:
         database_name = is_drop_database.group(1)
         baseTool.drop_database(database_name)
         return
@@ -63,9 +130,12 @@ def handle_command(command):
     
     if is_use_database:
         database_name = is_use_database.group(1)
-        current_database_name = baseTool.use_database(database_name)
-        if current_database_name != None:
-            print("Database changed : "+ current_database_name)
+        if database_name in di.keys() or "*" in di.keys(): 
+            current_database_name = baseTool.use_database(database_name)
+            if current_database_name:
+                print("Database changed : "+ current_database_name)
+        else:
+            print("ERROR : Access denied for user '%s' to database '%s'" % (username,database_name))
         return
 
     if is_show_tables:
@@ -75,21 +145,24 @@ def handle_command(command):
         baseTool.show_tables(current_database_name)
         return
 
-    if is_drop_table:
+    if is_drop_table and is_drop:
         if current_database_name == None:
             print("ERROR : No database selected")
             return
         table_name = is_drop_table.group(1)
-        baseTool.drop_table(current_database_name,table_name)
+        if table_name in di.get(current_database_name,[]) or "*" in (di.get(current_database_name,[]) or di.get("*",[])):
+            baseTool.drop_table(current_database_name,table_name)
+        else:
+            print("ERROR : Access denied for user '%s' to database '%s.%s'" % (username,current_database_name,table_name))
         return
 
         #    create table tabname(col1 type1 [not null] [primary key],col2 type2 [not null],..) 
   
-    if is_create_table:
+    if is_create_table and is_create:
         if current_database_name == None:
             print("ERROR : No database selected")
             return
-        # print("===============begin====================")
+        
         table_name = is_create_table.group(1)
         parms = is_create_table.group(2)
         baseTool.create_table(current_database_name,table_name,parms)
@@ -100,7 +173,10 @@ def handle_command(command):
             print("ERROR : No database selected")
             return
         table_name = is_describe_table.group(1)
-        baseTool.describe_table(current_database_name,table_name)
+        if table_name in di.get(current_database_name,[]) or "*" in (di.get(current_database_name,[]) or di.get("*",[])):
+            baseTool.describe_table(current_database_name,table_name)
+        else:
+            print("ERROR : Access denied for user '%s' to database '%s.%s'" % (username,current_database_name,table_name))
         return
 
     # 基本查询功能
@@ -109,32 +185,41 @@ def handle_command(command):
         if current_database_name == None:
             print("ERROR : No database selected")
             return
-
         table_name = is_alter_table_addcol.group(1)
         col_name = is_alter_table_addcol.group(2)
         parm = is_alter_table_addcol.group(3)
-        baseTool.alter_table_addcol(current_database_name,table_name,col_name,parm)
+        if table_name in di.get(current_database_name,[]) or "*" in (di.get(current_database_name,[]) or di.get("*",[])):
+            baseTool.alter_table_addcol(current_database_name,table_name,col_name,parm)
+        else:
+            print("ERROR : Access denied for user '%s' to database '%s.%s'" % (username,current_database_name,table_name))    
         return
 
-    if is_insert_into:
+    if is_insert_into and is_insert:
         if current_database_name == None:
             print("ERROR : No database selected")
             return
         table_name = is_insert_into.group(1)
         values = is_insert_into.group(2).split(",")
-        tableTool.insert_into(current_database_name,table_name,values)
+        if table_name in di.get(current_database_name,[]) or "*" in (di.get(current_database_name,[]) or di.get("*",[])):
+            tableTool.insert_into(current_database_name,table_name,values)
+        else:
+            print("ERROR : Access denied for user '%s' to database '%s.%s'" % (username,current_database_name,table_name))
+        
         return
 
-    if is_delete_from:
+    if is_delete_from and is_delete:
         if current_database_name == None:
             print("ERROR : No database selected")
             return
         table_name = is_delete_from.group(1)
         condition = is_delete_from.group(2)
-        tableTool.delete_from(current_database_name,table_name,condition)
+        if table_name in di.get(current_database_name,[]) or "*" in (di.get(current_database_name,[]) or di.get("*",[])):
+            tableTool.delete_from(current_database_name,table_name,condition)
+        else:
+            print("ERROR : Access denied for user '%s' to database '%s.%s'" % (username,current_database_name,table_name))
         return
     # update (\w+) set (\w+)=(\w+) where ([\s\S]+);
-    if is_update_table:
+    if is_update_table and is_update:
         if current_database_name == None:
             print("ERROR : No database selected")
             return
@@ -142,10 +227,14 @@ def handle_command(command):
         left = is_update_table.group(2).strip()
         right = is_update_table.group(3).strip()
         condition = is_update_table.group(4)
-        tableTool.update_table(current_database_name,table_name,left,right,condition)
+        if table_name in di.get(current_database_name,[]) or "*" in (di.get(current_database_name,[]) or di.get("*",[])):
+            tableTool.update_table(current_database_name,table_name,left,right,condition)
+        else:
+            print("ERROR : Access denied for user '%s' to database '%s.%s'" % (username,current_database_name,table_name))
+        
         return
     # select ([\s\S]+) from (\w+) where ([\s\S]+);
-    if is_select_from:
+    if is_select_from and is_select:
         if current_database_name == None:
             print("ERROR : No database selected")
             return
@@ -158,7 +247,11 @@ def handle_command(command):
             table_name = is_select_from_where_limit.group(2)
             condition = is_select_from_where_limit.group(3)
             limit = is_select_from_where_limit.group(4)
-            tableTool.select_from_where_limit(current_database_name,infos,table_name,condition,limit)
+            if table_name in di.get(current_database_name,[]) or "*" in (di.get(current_database_name,[]) or di.get("*",[])):
+                tableTool.select_from_where_limit(current_database_name,infos,table_name,condition,limit)
+            else:
+                print("ERROR : Access denied for user '%s' to database '%s.%s'" % (username,current_database_name,table_name))
+            
             return
         
         if is_select_from_where_order_by:
@@ -166,37 +259,52 @@ def handle_command(command):
             table_name = is_select_from_where_order_by.group(2)
             condition = is_select_from_where_order_by.group(3)
             order_key = is_select_from_where_order_by.group(4)
-            tableTool.select_from_where_order_by(current_database_name,infos,table_name,condition,order_key)
+            if table_name in di.get(current_database_name,[]) or "*" in (di.get(current_database_name,[]) or di.get("*",[])):
+                tableTool.select_from_where_order_by(current_database_name,infos,table_name,condition,order_key)
+            else:
+                print("ERROR : Access denied for user '%s' to database '%s.%s'" % (username,current_database_name,table_name))
+            
             return            
 
         table_name = is_select_from.group(2).strip()
         infos = is_select_from.group(1).strip()
         condition = is_select_from.group(3).strip()
-        tableTool.select_from(current_database_name,infos,table_name,condition)
+        if table_name in di.get(current_database_name,[]) or "*" in (di.get(current_database_name,[]) or di.get("*",[])):
+            tableTool.select_from(current_database_name,infos,table_name,condition)
+        else:
+            print("ERROR : Access denied for user '%s' to database '%s.%s'" % (username,current_database_name,table_name))
+        
         return
 
     # limit查询 select ([\s\S]+) from (\w+) where ([\s\S]+) limit ([\s\S]+);
-    if is_select_from_limit:
+    if is_select_from_limit and is_select:
         if current_database_name == None:
             print("ERROR : No database selected")
             return
         infos = is_select_from_limit.group(1)
         table_name = is_select_from_limit.group(2)
         limit = is_select_from_limit.group(3)
-        tableTool.select_from_limit(current_database_name,infos,table_name,limit)
+        if table_name in di.get(current_database_name,[]) or "*" in (di.get(current_database_name,[]) or di.get("*",[])):
+            tableTool.select_from_limit(current_database_name,infos,table_name,limit)
+        else:
+            print("ERROR : Access denied for user '%s' to database '%s.%s'" % (username,current_database_name,table_name))
         return
     # orderby 查询
-    if is_select_from_order_by:
+    if is_select_from_order_by and is_select:
         if current_database_name == None:
             print("ERROR : No database selected")
             return
         infos = is_select_from_order_by.group(1)
         table_name = is_select_from_order_by.group(2)
         order_key = is_select_from_order_by.group(3)
-        tableTool.select_from_order_by(current_database_name,infos,table_name,order_key)
+        if table_name in di.get(current_database_name,[]) or "*" in (di.get(current_database_name,[]) or di.get("*",[])):
+            tableTool.select_from_order_by(current_database_name,infos,table_name,order_key)
+        else:
+            print("ERROR : Access denied for user '%s' to database '%s.%s'" % (username,current_database_name,table_name))
+        
         return
     # group by 查询 select ([\s\S]+) from (\w+) group by ([\s\S]+); 
-    if is_select_from_group_by:
+    if is_select_from_group_by and is_select:
         if current_database_name == None:
             print("ERROR : No database selected")
             return
@@ -207,27 +315,99 @@ def handle_command(command):
             table_name = is_select_from_group_by_having.group(2)
             parms = is_select_from_group_by_having.group(3)
             condition = is_select_from_group_by_having.group(4)
-            tableTool.select_from_group_by_having(current_database_name,infos,table_name,parms,condition)
+            if table_name in di.get(current_database_name,[]) or "*" in (di.get(current_database_name,[]) or di.get("*",[])):
+                tableTool.select_from_group_by_having(current_database_name,infos,table_name,parms,condition)
+            else:
+                print("ERROR : Access denied for user '%s' to database '%s.%s'" % (username,current_database_name,table_name))
+            
             return
         infos = is_select_from_group_by.group(1)
         table_name = is_select_from_group_by.group(2)
         parms = is_select_from_group_by.group(3)
-        tableTool.select_from_group_by(current_database_name,infos,table_name,parms)
+        if table_name in di.get(current_database_name,[]) or "*" in (di.get(current_database_name,[]) or di.get("*",[])):
+            tableTool.select_from_group_by(current_database_name,infos,table_name,parms)
+        else:
+            print("ERROR : Access denied for user '%s' to database '%s.%s'" % (username,current_database_name,table_name))
+        
+        return
+
+    # 实现连接查询 select ([\s\S]+) from (\w+) right join (\w+) on ([\s\S]+)
+    if is_inner_join and is_select:
+        if current_database_name == None:
+            print("ERROR : No database selected")
+            return
+        infos = is_inner_join.group(1)
+        left_table = is_inner_join.group(2)
+        right_table = is_inner_join.group(3)
+        condition = is_inner_join.group(4)
+        if (left_table in di.get(current_database_name,[]) and right_table in di.get(current_database_name,[])) or "*" in (di.get(current_database_name,[]) or di.get("*",[])):
+            joinTool.inner_join(current_database_name,infos,left_table,right_table,condition)
+        else:
+            print("ERROR : Access denied for user '%s' to database '%s.%s'" % (username,current_database_name,left_table))
+            print("ERROR : Access denied for user '%s' to database '%s.%s'" % (username,current_database_name,right_table))
+        
+        return
+        
+    if is_left_join and is_select:
+        if current_database_name == None:
+            print("ERROR : No database selected")
+            return
+        infos = is_left_join.group(1)
+        left_table = is_left_join.group(2)
+        right_table = is_left_join.group(3)
+        condition = is_left_join.group(4)
+        if (left_table in di.get(current_database_name,[]) and right_table in di.get(current_database_name,[])) or "*" in (di.get(current_database_name,[]) or di.get("*",[])):
+            joinTool.left_join(current_database_name,infos,left_table,right_table,condition)
+        else:
+            print("ERROR : Access denied for user '%s' to database '%s.%s'" % (username,current_database_name,left_table))
+            print("ERROR : Access denied for user '%s' to database '%s.%s'" % (username,current_database_name,right_table))
+        
+        return
+    
+    if is_right_join and is_select:
+        if current_database_name == None:
+            print("ERROR : No database selected")
+            return
+        infos = is_right_join.group(1)
+        left_table = is_right_join.group(2)
+        right_table = is_right_join.group(3)
+        condition = is_right_join.group(4)
+        if (left_table in di.get(current_database_name,[]) and right_table in di.get(current_database_name,[])) or "*" in (di.get(current_database_name,[]) or di.get("*",[])):
+            joinTool.right_join(current_database_name,infos,left_table,right_table,condition)
+        else:
+            print("ERROR : Access denied for user '%s' to database '%s.%s'" % (username,current_database_name,left_table))
+            print("ERROR : Access denied for user '%s' to database '%s.%s'" % (username,current_database_name,right_table))
+        
         return
 
 def main():
     if (len(sys.argv) == 4) and (sys.argv[1] == "-u" or sys.argv[1] == "--user") and (sys.argv[3] == "-p" or sys.argv[3] == "--password"):
+        username = sys.argv[2]
         password = getpass("Enter your password>>>")
-        if password == "1":
-            print("Welcome to the MySQL monitor.  Commands end with ; or \g.")
-            while(True):
-                command = input("mysql>>>")
-                if(command == "exit;" or command == "quit;" or command == "\q;"):
-                    print("Bye")
-                    break         
-                while(not command.endswith(";")):
-                    command += input("    ->")
-                handle_command(command)
+
+        real_username = ""
+        real_password = ""
+        global userfile
+        with open(userfile,"r",newline="") as csvfile:
+            reader = csv.DictReader(csvfile,delimiter="|")
+            for row in reader:
+                if username == row["USERNAME"]:
+                    real_username = row["USERNAME"]
+                    real_password = row["PASSWORD"]
+            # 如果用户名不存在 
+            if real_password != password or real_username == "":
+                print("ERROR : Access denied for user '%s'" % username)
+                return
+
+        print("Welcome to the MySQL monitor.  Commands end with ; or \g.")
+        while(True):
+            command = input("mysql>>>")
+            while(not command.endswith(";")):
+                command += input("    ->")
+            if(command == "exit;" or command == "quit;" or command == "\q;"):
+                print("Bye")
+                break         
+            handle_command(real_username,command)
 
     else:
         show_usage()
